TGP-398 Create previous movement records page tests added and changes to Categorisation result page
