+++ src/test/resources/features/TraderGoodsProfile/01_ProfileSetup.feature
@@ -1,6 +1,6 @@
 @AllPipeline
 
-Feature: Trader Goods Profile - Profile setup journeys
+Feature: Trader Goods Profile - Profile setup and Previous movement records journey
 
   Scenario: As a UKIMS authorised trader I log on to the trader goods profile service with NIRMS and NIPHL
     Given I login to TGP Profile setup with EORI number GB123456789098
@@ -44,6 +44,10 @@ Feature: Trader Goods Profile - Profile setup journeys
     Then I should be on the 'Check Your Answers' page
     And My 'NIRMS Question' has the same value
     And I select continue
+    Then I should be on the 'Home' page
+    When I click on the View or update your goods records link
+    Then I should be on the 'Previous movement records' page
+    And I select the back link
     Then I should be on the 'Home' page
     And I click on the Sign out link
     Then I should be on the 'Signed Out' page
